Separating the universe classes into their own source and header file
